Colota list of codes em uma area comum
